requirements.txt
appnope==0.1.2
autopep8==1.5.5
backcall==0.2.0
blis==0.7.4
catalogue==1.0.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
cymem==2.0.5
decorator==4.4.2
deskew==0.10.3
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
funcy==1.15
future==0.18.2
gensim==3.8.3
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
inexactsearch==1.0.2
instagram-scraper==1.9.1
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.4
moviepy==1.0.3
murmurhash==1.0.5
networkx==2.5
nltk==3.5
numexpr==2.7.2
numpy==1.19.5
pandas==1.2.2
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
plac==1.1.3
preshed==3.0.5
proglog==0.1.9
prompt-toolkit==3.0.16
ptyprocess==0.7.0
pycodestyle==2.6.0
Pygments==2.8.0
pyLDAvis==3.2.2
pyparsing==2.4.7
pyspellchecker==0.5.6
pytesseract==0.3.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
regex==2020.11.13
requests==2.25.1
scikit-image==0.18.1
scipy==1.6.0
silpa-common==0.3
six==1.15.0
smart-open==4.2.0
soundex==1.1.3
spacy==2.3.5
spellchecker==0.4
srsly==1.0.5
tesseract==0.1.3
textblob==0.15.3
thinc==7.4.5
tifffile==2021.1.14
toml==0.10.2
tqdm==4.56.0
traitlets==5.0.5
urllib3==1.26.3
wasabi==0.8.1
wcwidth==0.2.5
wordninja==2.0.0
